Move OpRegistry functions to .cc file and move OpMaker to Op module
